Minimap now displays generated city, removed hunger
The game now loads in such a way that when the window loads, it inits
the playScreen. The playScreen creates the player, generates the map
(adding the player to the map, thus making the player object hold all
map info) and then sets the minimap overview to display the varios lots.
Additionally, I removed the hunger mixin and related functions as this
is not a game mechanic I want to be a part of this game.